+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The next release will require at least [Go 1.26].
 ### Added
 
 - Support testing of [Go 1.27]. (#8811)
+- Support `http/json` encoding in `go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploghttp`. (#8199)
 - Support `http/json` in `otlptracehttp` (#8273)
 - Add `Hasher` struct and methods in `go.opentelemetry.io/otel/attribute` to compute authoritative `Distinct` hashes incrementally for attribute filtering and deduplication. (#8598)